Check in scattered js changes, details below
This commit includes several changes to the scripts.js file. I got
distracted and wasn't commiting incrementally with a clear focus. In
this commit, I changed the basic math function to named arrow functions
in the helper section. I updated the formatting of the constant names
for the DOM elements. I moved the number/non-number check back to the
CALC_BUTTONS event listener and removed it from the handler functions. I
also began adding state management.