Develop state management further
Further development of state management has been the key to making the
logic for the calculator flow. This commit saves it in a working state
that still needs further development. The next step will require some
trial and error and I wanted to check in at a clean stopping point.